+++ R/Appender.R
@@ -523,7 +523,7 @@ AppenderTable <- R6::R6Class(
 #'   "test",
 #'   appenders = list(memory = AppenderDt$new()),
 #'   threshold = NA,
-#'   parent = NULL  # to prevent routing to root logger for this example
+#'   propagate = FALSE  # to prevent routing to root logger for this example
 #' )
 #' lg$debug("test")
 #' lg$error("test")