+++ R/get_logger.R
@@ -3,36 +3,41 @@ loggers <- new.env()
 
 
 
-#' Title
+#' Get/Create a Logger
 #'
-#' @param logger
-#' @param ...
+#' @param name
+#' @param class An [R6ClassGenerator][R6::R6] object. Usually `Logger` or `LoggerGlue`
+#'   are the only valid choices.
 #'
 #' @return
 #' @export
 #'
 #' @examples
+#'
+#' lg <- get_logger("log/ger/test")
+#' get_logger_glue("log/ger")
+#'
+#' lg
+#' lg$parent
+#'
 get_logger <- function(
   name,
-  ...
+  class = Logger
 ){
+  if (missing(name) || !length(name) || all(is_blank(name))){
+    return(lgr::lgr)
+  }
   nm_cur <- unlist(strsplit(name, "/", fixed = TRUE))
   name   <- paste(nm_cur, collapse = "/")
   res    <- get0(name, envir = loggers)
 
-  if (!is.null(res)){
+  if (inherits(res, class$classname)){
     return(res)
-  }
-
-  if (length(nm_cur) > 1){
-    parent <- get_logger(nm_cur[-length(nm_cur)])
-  } else {
-    parent <- get_logger("root")
   }
 
   assign(
     name,
-    Logger$new(name, parent = parent),
+    class$new(name),
     envir = loggers
   )
 
@@ -41,8 +46,43 @@ get_logger <- function(
 
 
 
+
+#' @rdname get_logger
+#' @export
+get_logger_glue <- function(
+  name
+){
+  if (is_virgin_Logger(name)){
+    get_logger(name = name, class = LoggerGlue)
+  } else {
+    warning("Logger '", name, "' already exists.")
+  }
+}
+
+
+
+
 exists_logger <- function(
   name
 ){
   inherits(get0(name, envir = loggers), "Logger")
+}
+
+
+
+
+is_virgin_Logger <- function(
+  x
+){
+  if (is.character(x)){
+    x <- get_logger(x)
+  } else {
+    assert(is_Logger(x))
+  }
+
+  !length(x$appenders) &
+  !length(x$filters) &
+  isTRUE(x$propagate) &
+  identical(x$exception_handler, default_exception_handler) &
+  is.null(x$.__enclos_env__$private$.threshold)
 }